refactor: previewRef optimization part 2
