+++ src/hooks/useSaveCanvasData.ts
@@ -1,4 +1,8 @@
-import { addDrawing } from '@/redux/history';
+import {
+  addDrawing,
+  selectActiveLayerIndex,
+  selectLayersList,
+} from '@/redux/history';
 import { useAppDispatch, useAppSelector } from '@/redux/hooks';
 import { combineCanvasImages } from '@/utils/combineCanvasImages';
 
@@ -6,12 +10,8 @@ export const useSaveCanvasData = (
   canvasElement: HTMLCanvasElement | null,
 ) => {
   const d = useAppDispatch();
-  const layersList = useAppSelector(
-    state => state.history.items?.[state.history.activeItemIndex]?.layersList,
-  ) ?? [];
-  const activeLayerIndex = useAppSelector(
-    state => state.history.items[state.history.activeItemIndex]?.activeLayerIndex,
-  );
+  const layersList = useAppSelector(selectLayersList);
+  const activeLayerIndex = useAppSelector(selectActiveLayerIndex);
 
   const saveCanvasData = (): void => {
     if (!canvasElement) return;
