+++ src/redux/history/selectors.ts
@@ -5,3 +5,13 @@ export const selectActiveLayer = (state: RootState) => {
   const activeLayerIndex = state.history.items[activeIndex]?.activeLayerIndex ?? 0;
   return state.history.items[activeIndex]?.layersList?.[activeLayerIndex];
 };
+
+export const selectActiveLayerIndex = (state: RootState) => {
+  const activeIndex = state.history.activeItemIndex;
+  return state.history.items[activeIndex]?.activeLayerIndex;
+};
+
+export const selectLayersList = (state: RootState) => {
+  const activeIndex = state.history.activeItemIndex;
+  return state.history.items[activeIndex]?.layersList ?? [];
+};
